+++ apx_core/src/http_signatures/verify.rs
@@ -98,6 +98,8 @@ pub struct HttpSignatureData {
     pub base: String, // recreated signature base
     pub signature: Vec<u8>,
     pub expires_at: DateTime<Utc>,
+    /// The authority of the target URI
+    pub authority: String,
     pub content_digest: Option<ContentDigest>,
 }
 
@@ -190,6 +192,11 @@ pub fn parse_http_signature_cavage(
     let signature_b64 = signature_parameters.get("signature")
         .ok_or(VerificationError::ParseError("signature is missing"))?;
     let signature = base64::decode(signature_b64)?;
+    let target_authority = request_headers.get("Host")
+        .ok_or(VerificationError::header_missing("Host"))?
+        .to_str()
+        .map_err(|_| VerificationError::header_value("Host"))?
+        .to_string();
     let created_at = if let Some(created_at) = signature_parameters.get("created") {
         let create_at_timestamp = created_at.parse()
             .map_err(|_| VerificationError::ParseError("invalid timestamp"))?;
@@ -257,6 +264,7 @@ pub fn parse_http_signature_cavage(
         base: signature_base,
         signature,
         expires_at,
+        authority: target_authority,
         content_digest: maybe_digest,
         is_rfc9421: false,
     };
@@ -264,7 +272,7 @@ pub fn parse_http_signature_cavage(
 }
 
 /// Parses RFC-9421 HTTP message signature  
-/// <https://datatracker.ietf.org/doc/html/rfc9421>
+/// <https://www.rfc-editor.org/rfc/rfc9421>
 pub fn parse_http_signature_rfc9421(
     request_method: &Method,
     request_uri: &Uri,
@@ -335,6 +343,9 @@ pub fn parse_http_signature_rfc9421(
     };
 
     check_required_components(&components, required_components)?;
+    let target_authority = request_uri.authority()
+        .ok_or(VerificationError::ParseError("incomplete request URI"))?
+        .to_string();
     let maybe_digest = get_content_digest(&components, request_headers)?;
 
     // Recreate signature base
@@ -362,16 +373,13 @@ pub fn parse_http_signature_rfc9421(
                 request_uri.path().to_owned()
             },
             "@query" => {
-                // https://datatracker.ietf.org/doc/html/rfc9421#name-query
+                // https://www.rfc-editor.org/rfc/rfc9421#name-query
                 let query = request_uri.query().unwrap_or_default();
                 format!("?{query}")
             },
             "@authority" => {
-                request_headers.get("host")
-                    .ok_or(VerificationError::header_missing("Host"))?
-                    .to_str()
-                    .map_err(|_| VerificationError::header_value("Host"))?
-                    .to_owned()
+                // https://www.rfc-editor.org/rfc/rfc9421#name-authority
+                target_authority.clone()
             },
             id if id.starts_with('@') => {
                 return Err(VerificationError::UnsupportedComponent(id.to_owned()));
@@ -398,6 +406,7 @@ pub fn parse_http_signature_rfc9421(
         base: signature_base,
         signature: signature_value,
         expires_at,
+        authority: target_authority,
         content_digest: maybe_digest,
         is_rfc9421: true,
     };
@@ -528,14 +537,10 @@ mod tests {
 
     #[test]
     fn test_parse_http_signature_rfc9421() {
-        // https://datatracker.ietf.org/doc/html/rfc9421#name-signing-a-request-using-ed2
+        // https://www.rfc-editor.org/rfc/rfc9421#name-signing-a-request-using-ed2
         let request_method = Method::POST;
-        let request_uri = "/foo?param=Value&Pet=dog".parse::<Uri>().unwrap();
+        let request_uri = "https://example.com/foo?param=Value&Pet=dog".parse::<Uri>().unwrap();
         let mut request_headers = HeaderMap::new();
-        request_headers.insert(
-            HeaderName::from_static("host"),
-            HeaderValue::from_static("example.com"),
-        );
         request_headers.insert(
             HeaderName::from_static("date"),
             HeaderValue::from_static("Tue, 20 Apr 2021 02:07:55 GMT"),