+++ mitra_activitypub/src/builders/add_note.rs
@@ -26,7 +26,7 @@ use crate::{
 #[derive(Serialize)]
 struct AddNote {
     #[serde(rename = "@context")]
-    context: Context,
+    _context: Context,
 
     #[serde(rename = "type")]
     activity_type: String,
@@ -51,7 +51,7 @@ fn build_add_note(
     let target_id = LocalActorCollection::Featured.of(&actor_id);
     let followers = LocalActorCollection::Followers.of(&actor_id);
     AddNote {
-        context: build_default_context(),
+        _context: build_default_context(),
         activity_type: ADD.to_string(),
         actor: actor_id,
         id: activity_id,