+++ mitra_activitypub/src/builders/delete_person.rs
@@ -20,7 +20,7 @@ use crate::{
 #[derive(Serialize)]
 struct DeletePerson {
     #[serde(rename = "@context")]
-    context: Context,
+    _context: Context,
 
     #[serde(rename = "type")]
     activity_type: String,
@@ -39,7 +39,7 @@ fn build_delete_person(
     let actor_id = local_actor_id(instance_uri, &user.profile.username);
     let activity_id = local_activity_id(instance_uri, DELETE, user.id);
     DeletePerson {
-        context: build_default_context(),
+        _context: build_default_context(),
         activity_type: DELETE.to_string(),
         id: activity_id,
         actor: actor_id.clone(),