+++ mitra_activitypub/src/builders/undo_announce.rs
@@ -29,7 +29,7 @@ use super::announce::{
 #[derive(Serialize)]
 pub struct UndoAnnounce {
     #[serde(rename = "@context")]
-    context: Context,
+    _context: Context,
 
     #[serde(rename = "type")]
     activity_type: String,
@@ -68,7 +68,7 @@ pub fn build_undo_announce(
         &recipient_id,
     );
     UndoAnnounce {
-        context: build_default_context(),
+        _context: build_default_context(),
         activity_type: UNDO.to_string(),
         id: activity_id,
         actor: actor_id,