+++ mitra_cli/src/commands/activitypub.rs
@@ -4,21 +4,30 @@ use anyhow::{anyhow, Error};
 use apx_sdk::{
     addresses::WebfingerAddress,
     authentication::verify_portable_object,
+    deliver::{send_object, DelivererError},
     fetch::FetchObjectOptions,
     utils::{get_core_type, CoreType},
 };
-use clap::Parser;
+use clap::{Parser, Subcommand};
 use serde_json::{Value as JsonValue};
+use uuid::Uuid;
 
 use mitra_activitypub::{
     agent::build_federation_agent,
     authentication::verify_signed_object,
+    authority::Authority,
+    builders::{
+        announce::build_relay_announce,
+        bite::build_bite,
+    },
+    deliverer::{Recipient, Sender},
+    identifiers::canonicalize_id,
     importers::{
-        fetch_any_object_with_context,
+        get_user_by_actor_id,
         import_activity,
+        import_actor,
         import_collection,
         import_object,
-        import_profile,
         import_replies,
         ApClient,
         CollectionItemType,
@@ -34,6 +43,8 @@ use mitra_models::{
         get_database_client,
         DatabaseConnectionPool,
     },
+    posts::queries::get_post_by_id,
+    profiles::queries::get_remote_profile_by_actor_id,
     users::{
         queries::{
             get_user_by_name,
@@ -109,7 +120,7 @@ impl ImportObject {
                 println!("post saved");
             },
             CoreType::Actor => {
-                import_profile(&ap_client, db_pool, object).await?;
+                import_actor(&ap_client, db_pool, object).await?;
                 println!("profile saved");
             },
             CoreType::Activity => {
@@ -205,31 +216,28 @@ impl FetchObject {
         } else {
             None
         };
-        let mut instance = config.instance();
+        let mut ap_client = ApClient::new_with_pool(config, db_pool).await?;
+        ap_client.as_user = maybe_user;
         match self.user_agent.as_deref() {
             Some("") =>
-                instance.user_agent = None,
+                ap_client.instance.user_agent = None,
             Some(user_agent) =>
-                instance.user_agent = Some(user_agent.to_owned()),
+                ap_client.instance.user_agent = Some(user_agent.to_owned()),
             None => (), // default
         };
-        let agent = build_federation_agent(
-            &instance,
-            maybe_user.as_ref(),
-        );
         let gateways = self.gateway
             .map(|gateway| vec![gateway])
             .unwrap_or_default();
         let mut context = FetcherContext::from(gateways);
+        let object_id = context.prepare_object_id(&self.object_id)?;
         let options = FetchObjectOptions {
-            skip_verification: self.skip_verification,
-            ..Default::default()
+            skip_content_type_verification: self.skip_verification,
         };
-        let object = fetch_any_object_with_context(
-            &agent,
-            &mut context,
-            &self.object_id,
+        let object = ap_client.fetch_object_raw(
+            &object_id,
             options,
+            self.skip_verification, // skip_authentication
+            vec![],
         ).await?;
         println!("{}", object);
         Ok(())
@@ -279,7 +287,7 @@ impl LoadPortableObject {
                 println!("object imported");
             },
             CoreType::Actor => {
-                import_profile(&ap_client, db_pool, object_json).await?;
+                import_actor(&ap_client, db_pool, object_json).await?;
                 println!("actor imported");
             },
             CoreType::Activity => {
@@ -287,6 +295,140 @@ impl LoadPortableObject {
                 println!("activity imported");
             },
             _ => return Err(anyhow!("unexpected object class")),
+        };
+        Ok(())
+    }
+}
+
+#[derive(Subcommand)]
+enum Activity {
+    /// mia:Bite activity
+    Bite {
+        /// Local username
+        sender: String,
+        /// Actor ID
+        recipient: String,
+    },
+    /// LitePub relay Announce activity
+    RelayAnnounce {
+        /// Internal post ID
+        post_id: Uuid,
+        /// Actor ID
+        recipient: String,
+    },
+}
+
+/// Create an activity with the specified parameters and print it
+#[derive(Parser)]
+pub struct CreateActivity {
+    #[clap(subcommand)]
+    activity: Activity,
+}
+
+impl CreateActivity {
+    pub async fn execute(
+        self,
+        config: &Config,
+        db_pool: &DatabaseConnectionPool,
+    ) -> Result<(), Error> {
+        let db_client = &**get_database_client(db_pool).await?;
+        let instance = config.instance();
+        let authority = Authority::from(&instance);
+        let activity = match self.activity {
+            Activity::Bite { sender, recipient } => {
+                let account = get_user_by_name(db_client, &sender).await?;
+                let target_profile =
+                    get_remote_profile_by_actor_id(db_client, &recipient).await?;
+                let bite = build_bite(
+                    &authority,
+                    &account.profile,
+                    target_profile.expect_actor_data(),
+                );
+                serde_json::to_value(bite)?
+            },
+            Activity::RelayAnnounce { post_id, recipient } => {
+                let post = get_post_by_id(db_client, post_id).await?;
+                let relay_actor = get_remote_profile_by_actor_id(db_client, &recipient).await?;
+                let announce = build_relay_announce(
+                    &authority,
+                    &post,
+                    relay_actor.expect_actor_data(),
+                );
+                serde_json::to_value(announce)?
+            },
+        };
+        println!("{activity}");
+        Ok(())
+    }
+}
+
+/// Send activity on behalf of a local user.
+///
+/// Activity will not be verified, and will not be added to the outbox.
+#[derive(Parser)]
+pub struct SendActivity {
+    /// JSON value
+    activity: String,
+    /// Actor ID
+    recipient: String,
+    /// Create RFC-9421 signature?
+    #[arg(long)]
+    rfc9421: bool,
+}
+
+impl SendActivity {
+    pub async fn execute(
+        self,
+        config: &Config,
+        db_pool: &DatabaseConnectionPool,
+    ) -> Result<(), Error> {
+        let instance = config.instance();
+        let authority = Authority::from(&instance);
+        let activity: JsonValue = serde_json::from_str(&self.activity)?;
+        let actor_id = activity["actor"].as_str()
+            .ok_or(Error::msg("'actor' property is missing"))?;
+        let canonical_actor_id = canonicalize_id(actor_id)?;
+        let account = get_user_by_actor_id(
+            db_client_await!(db_pool),
+            &authority,
+            &canonical_actor_id,
+        ).await?;
+        let recipient = get_remote_profile_by_actor_id(
+            db_client_await!(db_pool),
+            &self.recipient,
+        ).await?;
+        let recipient_inbox = Recipient
+            ::for_inbox(recipient.expect_actor_data())
+            .first()
+            .ok_or(Error::msg("recipient doesn't have an HTTP inbox"))?
+            .inbox
+            .clone();
+        let sender = Sender::from_user(instance.uri_str(), &account);
+        let mut agent = sender.into_agent(&instance);
+        if self.rfc9421 {
+            agent.rfc9421_enabled = true;
+        };
+        log::info!("sending activity to {recipient_inbox}");
+        match send_object(
+            &agent,
+            &recipient_inbox,
+            &activity,
+            &[],
+        ).await {
+            Ok(response) => {
+                println!("{}", response.status);
+                if !response.body.is_empty() {
+                    println!("{}", response.body);
+                };
+            },
+            Err(error) => {
+                println!("{error}");
+                if let DelivererError::HttpError(response) = error {
+                    if !response.body.is_empty() {
+                        println!("{}", response.body);
+                    };
+                };
+            },
         };
         Ok(())
     }